Black Hole Gun
How to use?
Put the BlackHoleGun script on the grapplinghook prefab and deactivate
the grapplinghook script.
Put the GravityObject script on every object that should be affected by
the Black Hole’s gravitation force (so put it also on the player!).

Still to do:
Spawnpoint of the orb is not exactly at the centerpoint of the gun
We need a different gun model for this weapon
We need a model/sprite/effect for the black hole orb
Maybe add explosion once the orb is removed